invitation CRUD

+++ app/api/animals/[id]/route.ts
@@ -1,4 +1,4 @@
-import { deleteAnimal, editAnimal, getById } from "@/actions/animalActions";
+import { deleteAnimal, editAnimal, getAnimalById, getById } from "@/actions/animalActions";
 import { request } from "http";
 import { NextRequest, NextResponse } from "next/server"
 
@@ -16,26 +16,39 @@ export async function GET(request : NextRequest, context : { params: { id: strin
   }
 
 
-export const PUT = async (request: NextRequest, context: { params: { id: string } }) => {
+  export const PUT = async (request: NextRequest, context: { params: { id: string } }) => {
     const id = parseInt(context.params.id);
 
     try {
         const body = await request.json();
         const { breed, species, name, age, vaccinated, trained, friendly, available, description, image } = body;
-        await editAnimal(id, breed, species, name, age, vaccinated, trained, friendly, available, description, image);
 
+        // Vérifiez l'existence de l'animal
+        const existingAnimal = await getAnimalById(id);
+        if (!existingAnimal) {
+            return NextResponse.json({ message: 'Animal not found' }, { status: 404 });
+        }
+
+        // Mettez à jour l'animal s'il existe
+        await editAnimal(id, breed, species, name, age, vaccinated, trained, friendly, available, description, image);
+        
         return NextResponse.json({ message: 'Animal updated successfully' }, { status: 200 });
     } catch (error: any) {
         return NextResponse.json({ message: 'Error updating animal', error }, { status: 500 });
-
     }
-}
-
+};
 
 export const DELETE = async (request: NextRequest, context: { params: { id: string } }) => {
     const id = parseInt(context.params.id);
 
     try {
+        
+        const existingAnimal = await getAnimalById(id);
+        if (!existingAnimal) {
+            return NextResponse.json({ message: 'Animal not found' }, { status: 404 });
+        }
+
+
         await deleteAnimal(id);
         return NextResponse.json({ message: 'Animal deleted successfully' }, { status: 200 });
